Exception handler

+++ src/main/java/com/githunb/ailton78/dscommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.githunb.ailton78.dscommerce.services;
 import com.githunb.ailton78.dscommerce.dto.ProductDTO;
 import com.githunb.ailton78.dscommerce.entities.Product;
 import com.githunb.ailton78.dscommerce.repositories.ProductRepository;
+import com.githunb.ailton78.dscommerce.services.exceptions.DatabaseException;
+import com.githunb.ailton78.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 
@@ -17,13 +22,14 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).orElseThrow(
+                ()->new ResourceNotFoundException("Recurso não encontrado"));
     return new ProductDTO(product);
     }
     @Transactional(readOnly = true)
     public Page<ProductDTO> findAll(Pageable pageable){
        Page<Product> result= repository.findAll(pageable);
-        return result.map(x -> new ProductDTO(x));
+        return result.map(ProductDTO::new);
     }
     @Transactional
     public ProductDTO insert(ProductDTO dto){
@@ -35,15 +41,28 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-       Product entity = repository.getReferenceById(id);
-       copyDtoToEntity(dto, entity);
-       entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }catch (EntityNotFoundException e){
+            throw  new ResourceNotFoundException("Recurso não encontrado");
+        }
+
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw  new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
+
     }
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
         entity.setName(dto.getName());
